refactor: 내부 enum 클래스로 책임을 한곳으로 옮긴다
